refactor(user): move user, group and auth interfaces to model files
