+++ src/app/core/models/auth.model.ts
@@ -1,0 +1,13 @@
+export interface User {
+  id: number;
+  email: string;
+  name?: string;
+}
+
+export interface LoginResponse {
+  success: boolean;
+  data: {
+    user: User;
+    token: string;
+  };
+}
